+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Release v0.17.0
+* The `utf8_text` library is now deprecated. Users of this library should migrate to
+  `Core.String.Utf8`.
+
 ## Release v0.16.0
 
-* Added [Text_block.span_banner] to indicate boundaries of text content
+* Added `Text_block.span_banner` to indicate boundaries of text content